ai-data-platform-package/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-data-platform"
version = "0.1.0"
description = "A reusable, AI-driven backend for querying third-party platforms."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "DataBridge AI"},
]
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "stripe>=7.0.0",
    "zenpy>=2.0.0",
    "cryptography>=41.0.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://example.com"

[tool.setuptools.packages.find]
where = ["src"]
